Add new gen_stream behaviour for efficiently consuming serial streams
Introduce a new behaviour for serially processing text files, generating
lazy or infinite streams or providing a serial interface for a binary.
Implement the behaviour functionality using a gen_server while internally
managing multiple processes to load the stream.  Stripe stream data
across the worker processes to enable overlapped I/O when reading from
the source. Impose a simple API for consuming the stream serially, a block
at a time.

The initial arguments for the gen_stream specify the number of processes
used, the number of data chunks per process, the size of each chunk a
blocking factor for reading multiple chunks at a time and an optional
function parameter to transform the data while buffering it.  The serial
stream delivers the (potentially transformed) data from the internal
process buffers as it is requested by the calling application.