Add flag-based setting for 'erts_de_busy_limit' constant
A prior patch to change the 'erts_de_busy_limit' constant, based on an
environment variable, was rejected by the Ericsson OTP team because
environment variables are no longer used for modifying the VM's
behavior.

Families of flags like "+s" are devoted to the scheduler (except for
"+sss"), but it wasn't obvious to me where a "+"-style flag for
'erts_de_busy_limit' should go.  So, this patch creates a new family
of flags, prefixed with "+z", for miscellaneous VM settings.

The full flag name is "+zerts_de_busy_limit", which goes against the
tradition of using "+Qmbl" for adjusting some kind of "mumble mumble"
value but "mumble mumble" is obviously too long to be useful.  If
"+zerts_de_busy_limit" is too long, then I'll leave it to the OTP team
to create a tradition-fitting, impossible-to-remember abbreviation to
use instead.

I don't a functional Java compiler with the suitable Java compliance
level, so compiling & running the test suite isn't an option for me
yet.  I've made the attempt to write a test, but I haven't executed it
to know if it works.  erlang:system_info(erts_de_busy_limit) was added
to make a unit test possible.

The documentation changes are terse but present.  Given the lack of
Java above, my doc-building toolchain is broken, so I don't know if
the formatting is correct, sorry.