Fix crash when calling functions in a module with an on_load attribute from a native module
The fix consists in adding a new opcode for hipe stub calls.

It is based on the following understanding of the loader: when a module is loaded, the export entries are first allocated as stubs to call_error_handler, which loads the function or generates an error. Then, the entries are updated with the address to the function, except when the module has an on_load attribute. Export entries of modules with an on_load attribute are not made callable immediatly. Instead, the function address is saved and the on_load function is called. If this succeeds, the entry is finally fixed. The issue is that hipe can generate a stub with a pointer to the call_error_handler opcode set the loader before on_load function is called.

The fix consists in updating the call_error_handler opcode with a new opcode that simply calls the loaded function.