Optimize string table generation for beams
Original code was based on a naive string search operation (which was actually
duplicated from string:str/2) plus expensive erlang:++ concatenation. This
could prove extremely slow when compiling modules with a lot of (bit)strings
(several minutes for a single module with thousands of strings on modern
hardware). Besides old code was wrongly typed: the pool was declared as a
[string()] while it was a string().

New code is based on Boyer–Moore algorithm through binary:match/2 native
implementation. All strings are converted to binary (with old code, the pool
was eventually converted to binary when output as a beam chunk). A new test
asserts the non-regression.